Description of problem extended.
Removed Windows example parts, since this specific case applies only on
linux setup. But in general, this same failure can happen in any OS, and
it is only this example, that does not make sense on other OSes.

Description of problem was given more details, but still "work in
progress" very much.